+++ perimeterx/px_httpc.py
@@ -1,8 +1,18 @@
 import time
+
 import requests
 
 
 def send(full_url, body, headers, config, method):
+    """
+    Returns the appropriate response parameters according to blocking parameters
+    :param string full_url:
+    :param string body:
+    :param dict headers:
+    :param PxConfig config:
+    :param string method:
+    :return response: requests.response type
+    """
     logger = config.logger
     try:
         start = time.time()
@@ -14,9 +24,10 @@ def send(full_url, body, headers, config, method):
         if response.status_code >= 400:
             logger.debug('PerimeterX server call failed')
             return False
-
-        logger.debug('PerimeterX server call took ' + str(time.time() - start) + 'ms')
+        finish = time.time()
+        request_time = finish - start
+        logger.debug('PerimeterX server call took {} ms'.format(request_time * 1000))
         return response
-    except requests.exceptions.RequestException as e:
-        logger.debug('Received RequestException, message: ' + e.message)
-        return False
+    except requests.exceptions.RequestException as err:
+        logger.debug('Received RequestException. Error: {}'.format(err))
+        raise err